Move OverallDeployment function to its own file
Need to reuse all these so better move them to a new file.
